Move TestApiClient to utils app and use it in api integration test as well
